# src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Endpoint of EX Bank
  version: v1
paths:
...
  /users:
    get:
      tags:
        - users
      security:
        - bearerAuth: [ ]
      summary: Returns a list of users.
      description: Optional extended description in CommonMark or HTML
      responses:
        "200":
          description: A JSON array of user
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
        "204":
          description: no users found
    post:
      tags:
        - users
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'
      responses:
        "201":
          description: user created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "409":
          description: user already exist
  /users/{id}:
    get:
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric Id of User
      responses:
        "200":
          description: user exists
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "204":
          description: user not found
    put:
      summary: Update User
      operationId: updateUser
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric Id of User
      requestBody:
        description: Updated user object
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/User'

      responses:
        "201":
          description: User is updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "204":
          description: user not found
    delete:
      summary: Delete User
      operationId: deleteUser
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric Id of User
      responses:
        "200":
          description: user deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "204":
          description: user not found
        "226":
            description: user has accounts
...
  /users/{id}/accounts:
    post:
      summary: Create Account
      operationId: createAccount
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric Id of User
      requestBody:
        description: Account
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Account'
      responses:
        "201":
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
    get:
      summary: Returns a list of account.
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: Numeric Id of User
      description: Optional extended description in CommonMark or HTML
      responses:
        "200":
          description: A JSON array of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Account"
        "204":
          description: no accounts
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      required:
        - firstName
        - lastName
        - dateOfBirth
        - role
      type: object
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        accounts:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/Account'
        firstName:
          type: string
          minLength: 1
          maxLength: 20
          pattern: '\p{L}+'
          example: John
        lastName:
          type: string
          minLength: 1
          maxLength: 20
          pattern: '\p{L}+'
          example: James
        dateOfBirth:
          type: string
          format: date
          example: '1995-11-27'
        email:
          type: string
          example: example@mail.com
        password:
          type: string
          example: '12345'
          writeOnly: true
        phone:
          type: string
          example: '12345'
        userStatus:
          type: integer
          description: User Status
          example: 1
        role:
          $ref: '#/components/schemas/Role'
        dateOfCreation:
          type: string
          format: date-time
          readOnly: true

      xml:
        name: user
    Account:
      type: object
      required:
        - userId
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        userId:
          type: integer
          format: int64
          readOnly: true
        transactions:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/Transaction'
        iban:
          type: string
          readOnly: true
        accountNumber:
          type: integer
          pattern: '[0-9]{8}'
          readOnly: true
        amount:
          type: float
          readOnly: true
        currency:
          type: string
          example: 'EUR'
        isMain:
          type: boolean
          example: true
        dateOfCreation:
          type: string
          format: date-time
          readOnly: true
    Role:
      type: string
      enum:
        - user
        - bank employee
        - admin
    Transaction:
      required:
        - sourceIban
        - destinationName
        - destinationIban
        - amount
        - currency
        - purposeOfUse
      properties:
        id:
          type: integer
          format: int64
          readOnly: true
        accounts:
          type: array
          readOnly: true
          items:
            $ref: '#/components/schemas/Account'
        sourceName:
          type: string
          readOnly: true
        sourceIban:
          type: string
          pattern: '\p{L}{2}[0-9]{16}'
          example: 'DE9111111111000001'
        destinationName:
          type: string
          pattern: '[a-zA-Z\s]+'
          example: 'John Adams'
        destinationIban:
          type: string
          pattern: '\p{L}{2}[0-9]{16}'
          example: 'DE9111111111000001'
        destinationBIC:
          type: string
          readOnly: true
        amount:
          type: float
          example: 570.00
        currency:
          type: string
          example: 'EUR'
        purposeOfUse:
          type: string
          example: 'Miete Januar'
        dateOfTransaction:
          type: string
          format: date-time
          example: '2024-12-02T18:00:00+01:00'
        dateOfCreation:
          type: string
          format: date-time
          readOnly: true
...